// client-monorepo/libs/applets/insurance/src/lib/app/features/vehicle/features/third-party/data-access/enums/third-party-keys.enum.ts
export enum ThirdPartyKeysEnum {
  CarType = 'carType',
  CarUsage = 'carUsage',
  CarBrand = 'carBrand',
  CarModel = 'carModel',
  CarYear = 'buildYear',
  ExInsurer = 'exInsurer',
  ExInsurerStart = 'exInsurerStart',
  ExInsurerEnd = 'exInsurerEnd',
  EX_INSURER_POLICY_NUMBER = 'exInsurerPolicyNumber',
  NoPlate = 'np',
  Sanhab = 'isSan',
  Discount = 'discount',
  ThirdPartyDiscount = 'thirdPartyDiscount',
  DriverDiscount = 'driverDiscount',
  OwnershipChange = 'ownershipChange',
  Claim = 'claim',
  PropertyDamage = 'propertyDamage',
  DriverDamage = 'driverDamage',
  HealthDamage = 'healthDamage',
  PLPCoverageRateInput = 'PLPCoverageRateInput',
  PLPDurationInput = 'PLPDurationInput',
  UniqueCode = 'order',
  SessionId = 'session',
  Interval = 'i',
  PaymentMethod = 'paymentMethod',
  ProviderId = 'providerId',
  CarDataInvalid = 'cDI',
  InsurerDataInvalid = 'iDI',
  UrlGoToPayment = 'r',
  Referrer = 'referrer',
  IsPlatesListOpen = 'isPlatesListOpen',
  NoCheck = 'noCheck',
  FormId = 'id',
  SelectedPLPCardCompanyId = 'selectedPLPCardCompanyId'
}