// client-monorepo/libs/applets/insurance/src/lib/app/data-access/enums/insurance-urls.enum.ts
// @ToDo merge this enum with insurance-product-type enum and product-type-base-url
export enum InsuranceUrlsEnum {
  PolicyDetail = 'policy/detail',
  VehicleThirdParty = 'vehicle/third-party',
  VehicleThirdPartyMotor = 'vehicle/third-party-motor',
  PolicyList = 'policy/list',
  Body = 'vehicle/body',
  Faq = 'faq',
  TermsAndCondition = '/terms-and-condition',
  Equipment = 'equipment/used',
  EquipmentRenewal = 'equipment/renewal',
  Subscription = 'subscription'
}